requirements.txt
pypdf>=3.0.0
PyMuPDF>=1.20.0
Pillow>=10.0.0